+++ guava-rxjava/src/test/java/net/javacrumbs/futureconverter/guavarx/ToListenableFutureConverterTest.java
@@ -17,18 +17,18 @@ package net.javacrumbs.futureconverter.guavarx;
 
 import com.google.common.util.concurrent.ListenableFuture;
 import net.javacrumbs.futureconverter.common.test.guava.GuavaConvertedFutureTestHelper;
-import net.javacrumbs.futureconverter.common.test.rxjava.AbstractObservableToFutureConverterTest;
+import net.javacrumbs.futureconverter.common.test.rxjava.AbstractSingleToFutureConverterTest;
 import rx.Single;
 
-public class ToListenableFutureConverterTest extends AbstractObservableToFutureConverterTest<ListenableFuture<String>> {
+public class ToListenableFutureConverterTest extends AbstractSingleToFutureConverterTest<ListenableFuture<String>> {
 
     public ToListenableFutureConverterTest() {
         super(new GuavaConvertedFutureTestHelper());
     }
 
     @Override
-    protected ListenableFuture<String> toFuture(Single<String> observable) {
-        return FutureConverter.toListenableFuture(observable);
+    protected ListenableFuture<String> toFuture(Single<String> single) {
+        return FutureConverter.toListenableFuture(single);
     }
 
     @Override