+++ rxjava-java8/src/test/java/net/javacrumbs/futureconverter/java8rx/ToSingleConverterTest.java
@@ -13,26 +13,26 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package net.javacrumbs.futureconverter.guavarx;
+package net.javacrumbs.futureconverter.java8rx;
 
-import com.google.common.util.concurrent.ListenableFuture;
-import net.javacrumbs.futureconverter.common.test.guava.GuavaOriginalFutureTestHelper;
-import net.javacrumbs.futureconverter.common.test.rxjava.AbstractFutureToObservableConverterTest;
+import net.javacrumbs.futureconverter.common.test.java8.Java8OriginalFutureTestHelper;
+import net.javacrumbs.futureconverter.common.test.rxjava.AbstractFutureToSingleConverterTest;
 import rx.Single;
 
-public class ToObservableConverterTest extends AbstractFutureToObservableConverterTest<ListenableFuture<String>> {
+import java.util.concurrent.CompletableFuture;
 
-    public ToObservableConverterTest() {
-        super(new GuavaOriginalFutureTestHelper());
+public class ToSingleConverterTest extends AbstractFutureToSingleConverterTest<CompletableFuture<String>> {
+    public ToSingleConverterTest() {
+        super(new Java8OriginalFutureTestHelper());
     }
 
     @Override
-    protected Single<String> toSingle(ListenableFuture<String> future) {
+    protected Single<String> toSingle(CompletableFuture<String> future) {
         return FutureConverter.toSingle(future);
     }
 
     @Override
-    protected ListenableFuture<String> toFuture(Single<String> observable) {
-        return FutureConverter.toListenableFuture(observable);
+    protected CompletableFuture<String> toFuture(Single<String> single) {
+        return FutureConverter.toCompletableFuture(single);
     }
 }